Allow building multiple wallets (#209)
