+++ src/app/components/seedPhraseInput/index.tsx
@@ -68,7 +68,7 @@ const SeedWordInput = React.forwardRef<HTMLInputElement, SeedWordInputProps>(
     const handlePasteInput = (e: React.ClipboardEvent<HTMLInputElement>) => {
       e.preventDefault();
 
-      if (DEV_MODE) {
+      if (DEV_MODE || process.env.WALLET_LABEL) {
         const { clipboardData } = e;
         const pastedText = clipboardData.getData('text');
         handlePaste(pastedText);
